+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Go to/peek definition of messages.
 - Format centered messages, keywords, symbol definitions, tasks and comments.
 - Support for quests in the workspace: autocomplete, hover, go to/peek definition and references.
+- Import modules for actions/conditions autocomplete and hover.
 
 ### Improved
 - Detection of language symbols (prefix: `%`) and quest symbols with prefix `=` or `==`.